repocheck: avoid hardcoding remotes
Tried this way before, but it felt clunky to use methods from go-git.
But hardcoding remotes weren't that good either. Worked for me in most places that I wanted to check.
So going the clunky way now.
Still not nice, that the remote is covered behind a number, but couldn't figure out on how just to get the current remote name.